refactor(tests): create dedicated classes for each endpoint
