+++ src/test/java/ch/psi/ord/api/ExportTestIT.java
@@ -1,0 +1,6 @@
+package ch.psi.ord.api;
+
+import io.quarkus.test.junit.QuarkusIntegrationTest;
+
+@QuarkusIntegrationTest
+public class ExportTestIT extends ExportTest {}
