+++ src/test/java/ch/psi/ord/api/ImportTestIT.java
@@ -1,0 +1,6 @@
+package ch.psi.ord.api;
+
+import io.quarkus.test.junit.QuarkusIntegrationTest;
+
+@QuarkusIntegrationTest
+public class ImportTestIT extends ImportTest {}
